% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgt.R
\name{cgt}
\alias{cgt}
\title{Title}
\usage{
cgt(
  core_threshold = 0.95,
  rare_threshold = 0.1,
  completeness,
  observations,
  quantiles = c(0.05, 0.25, 0.5),
  plot = TRUE,
  betap1 = 1,
  betap2 = 1,
  n_posterior_samples = 10000,
  model_object
)
}
\arguments{
\item{core_threshold}{The true frequency above which a gene is considered a core gene (numeric)}

\item{completeness}{A numeric vector of completeness values for each genome sample from CheckM, must be between 0 and 1}

\item{observations}{A vector of the number of observations of each gene, must be between 0 and the number of genome samples}

\item{n.cores}{The number cores to use for parallel calculation, useful when there are >1000 genome samples}
}
\description{
Title
}
\examples{
cgt(core_threshold = 0.95,
    completeness = runif(20, 0.95, 1),
    quantiles = c(0.3, 0.5),
    observations = rbinom(140, size = 20, prob = 0.90))

}
